\name{dateList}
\alias{dateList}
\docType{data}
\title{
Date list.
}
\description{
Date list.
}
\usage{data("dateList")}
\format{
  The format is:
 chr [1:1827] "2009-01-01" "2009-01-02" "2009-01-03" "2009-01-04" ...
}
\source{
Vietnam stock market.
}
\references{
See the report.
}
\examples{
data(dateList)
}
