\name{jStockList}
\alias{jStockList}
\title{
Stocks List in Vietnam stock market.
}
\description{
Provide a list of Vietnamese stocks.

See the report: Value at Risk.<researchgate.net>
}
\usage{
jStockList()
}
\value{
A list of Vietnamese stocks
}
\author{
Hung Vu
}
\note{
viet-hung.vu@jvn.edu.vn
}


\seealso{
https://www.researchgate.net/profile/Vu_Hung4
}
\examples{
jStockList()
}
\keyword{ jVaR }
\keyword{ jPrice }
\keyword{ jReturn }
