\name{jvnVaR-package}
\alias{jvnVaR-package}
\alias{jvnVaR}
\docType{package}
\title{
Value at risk package.
}
\description{
Provide many method to compute, predict and back-test VaR.

More about what it does, see the report: Value at Risk.<researchgate.net>
}
\details{
\tabular{ll}{
Package: \tab jvnVaR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-10\cr
License: \tab GPL-3\cr
}
Using command 'jListFunctions()' to know its useful functions.
}
\author{
Hung Vu

Maintainer: Hung Vu <viethungttbk@gmail.com>
}
\references{
See the report: Value at Risk.<Researchgate.net>
}
\keyword{ jVaR }
\seealso{
https://www.researchgate.net/profile/Vu_Hung4
}
