\name{stockList}
\alias{stockList}
\docType{data}
\title{
Stock List.
}
\description{
Stock List.
}
\usage{data("stockList")}
\format{
  The format is:
 chr [1:691] "AAA" "AAM" "ABT" "ACB" "ACC" "ACL" "ADC" ...
}

\source{
http://www.cophieu68.vn/export.php
https://www.vndirect.com.vn/portal/thong-ke-thi-truong-chung-khoan/lich-su-gia.shtml

}
\examples{
data(stockList)
}

